// src/types/enums/vehicleEnums.ts
export enum VehicleUtilizationStyle {
  Unknown = 0,
  PrivateCar = 1,
  Taxi = 2,
  RouteBasedMinibus = 3,
  MediumBus = 4,
  LargeBus = 5,
  PickupTruck = 6,
  ClosedBedPickup = 7,
  Truck = 8,
  ConstructionMachinery = 9,
  Tractor = 10,
  Trailer = 11,
  Motorcycle = 12,
  Tanker = 13,
  TowTruck = 14,
  MotorizedCaravan = 15,
  TowableCaravan = 16,
  AgriculturalMachineExcludingTractor = 17,
  OpenBodyTruck = 18,
  RentalCar = 19,
  ArmoredVehicle = 20,
  MinibusSharedTaxi = 21,
  Jeep = 22,
  JeepSAV = 23,
  JeepSUV = 24,
  JeepRental = 25,
  JeepTaxi = 26,
  Ambulance = 27,
  FirefighterCar = 28,
  Hearse = 29,
  ChauffeuredRentalCar = 30,
  OperationalRental = 31,
  PrivateMinibus = 32,
  RouteMinibus = 33,
  ServiceMinibus = 34,
  CompanyMinibus = 35,
  RentalMinibus = 36,
  AmbulanceMinibus = 37,
  MinibusBroadcastingVehicle = 38,
  MinibusArmoredTransport = 39,
  SmallBus1535Passengers = 40,
  SmallBusService = 41,
  SmallBusCity = 42,
  SmallBusRoute = 43,
  LargeBus36Plus = 44,
  DumpTruck = 45,
  RefrigeratedTruck = 46,
  TruckWithConcreteMixer = 47,
  SiloTruck = 48,
  TruckWithConcretePump = 49,
  RockTruck = 50,
  TruckWithCrane = 51,
  HeavyMachinery = 52,
  Excavator = 53,
  Loader = 54,
  Bulldozer = 55,
  Scraper = 56,
  Grader = 57,
  RoadRoller = 58,
  MobileCrane = 59,
  IndoorForklift = 60,
  OutdoorForklift = 61,
  MobileCompressor = 62,
  MobilePump = 63,
  MobileWeldingMachine = 64,
  CombineHarvester = 65,
  TankerAcidCarrier = 66,
  TankerWaterFuelCarrier = 67,
  TankerExplosiveFlammable = 68,
  TowTruckTractor = 69,
  TowTruckTanker = 70,
  PanelGlassVanMinubus = 71,
}

export enum VehicleFuelType {
  Gasoline = 1,
  Diesel = 2,
  Lpg = 3,
  Electric = 4,
  LpgGasoline = 5,
}